Backend/requirements.txt
# FastAPI and web framework dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# PDF processing
PyMuPDF==1.24.10
pytesseract==0.3.10
Pillow==10.1.0

# Text processing and NLP
nltk==3.8.1
scikit-learn==1.3.2
numpy==1.26.4

# Machine Learning and Transformers (optional: uncomment if using abstractive)
torch>=2.2.0
transformers>=4.35.0
tokenizers>=0.14.0

# Data processing
pandas==2.1.3

# HTTP client
httpx==0.25.2

# Environment and configuration
python-dotenv==1.0.0

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0

# Optional: For better performance
# accelerate==0.24.1
# bitsandbytes==0.41.3